[ALSA-OSS] Always initialize pointers before use
This patch makes sure that the initialize() is always called
before any pointer to function is used. Somehow it happens
that the __attribute__((constructor)) is not always sufficient -
other libs may have that attribute either, and may call fopen()
from within their constructors, which leads to a SIGSEGV, that
this patch fixes. Perhaps the __attribute__((constructor)) is no
longer needed with that patch at all.

Signed-off-by: Stas Sergeev <stsp@aknet.ru>